chess_arena/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pydantic>=2.6
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-chess>=1.999
glicko2>=2.1
httpx>=0.27
websockets>=12.0
pyside6>=6.6
